[CSES] - Weird Algorithm type correction
Due to the size of the input values, the `int` data type was
experiencing an overflow and resulted in infinite loops or invalid
values from the tests. In order to resolve the invalid test results and
data overflow, the `int` data type will be updated to `unsigned long
int`.

- Update `weirdInput` data type from `int` to `unsigned long int`

+++ CSES/Introductory_Problems/Weird_Algorithm/weirdalgorithm.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 
 int main() {
-  int weirdInput = 0;
+  unsigned long int weirdInput = 0;
 
   cin >> weirdInput;
